books/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-bootstrap-form==3.4
django-filter==2.4.0
djangorestframework==3.12.4
idna==3.2
numpy==1.21.2
pandas==1.3.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6